pyproject.toml
[project]
name = "nt-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.43.2",
    "bear-lake>=0.1.5",
    "fastapi[standard]>=0.128.0",
    "numpy>=2.4.0",
    "pandas-market-calendars>=5.2.4",
    "polars>=1.37.1",
    "python-dotenv>=1.2.1",
]
